+++ src/private/release/GetLocalModulePathErrorMessage.ps1
@@ -1,0 +1,11 @@
+function Get-LocalModulePathErrorMessage {
+    param(
+        [Parameter(Mandatory)][string]$MatchPattern
+    )
+
+    if ($IsWindows) {
+        return "No windows module path matching $MatchPattern found"
+    }
+
+    return "No macOS/Linux module path matching $MatchPattern found in PSModulePath."
+}