+++ src/private/release/GetLocalModulePathPattern.ps1
@@ -1,0 +1,7 @@
+function Get-LocalModulePathPattern {
+    if ($IsWindows) {
+        return '\\Documents\\PowerShell\\Modules'
+    }
+
+    return '/\.local/share/powershell/Modules$'
+}